// blockchain/typechain-types/GameLogic.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import type {
  BaseContract,
  BigNumberish,
  BytesLike,
  FunctionFragment,
  Result,
  Interface,
  AddressLike,
  ContractRunner,
  ContractMethod,
  Listener,
} from "ethers";
import type {
  TypedContractEvent,
  TypedDeferredTopicFilter,
  TypedEventLog,
  TypedListener,
  TypedContractMethod,
} from "./common";

export interface GameLogicInterface extends Interface {
  getFunction(
    nameOrSignature: "calculatePayouts" | "getRandom" | "getRandomNumbers"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "calculatePayouts",
    values: [BigNumberish, AddressLike[], BigNumberish[], BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getRandom",
    values: [BigNumberish, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getRandomNumbers",
    values: [BigNumberish, BigNumberish, BigNumberish]
  ): string;

  decodeFunctionResult(
    functionFragment: "calculatePayouts",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "getRandom", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "getRandomNumbers",
    data: BytesLike
  ): Result;
}

export interface GameLogic extends BaseContract {
  connect(runner?: ContractRunner | null): GameLogic;
  waitForDeployment(): Promise<this>;

  interface: GameLogicInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  calculatePayouts: TypedContractMethod<
    [
      balance: BigNumberish,
      wallets: AddressLike[],
      percents: BigNumberish[],
      ownerFeePercent: BigNumberish
    ],
    [bigint[]],
    "view"
  >;

  getRandom: TypedContractMethod<
    [a: BigNumberish, b: BigNumberish],
    [bigint],
    "view"
  >;

  getRandomNumbers: TypedContractMethod<
    [a: BigNumberish, b: BigNumberish, count: BigNumberish],
    [bigint[]],
    "view"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "calculatePayouts"
  ): TypedContractMethod<
    [
      balance: BigNumberish,
      wallets: AddressLike[],
      percents: BigNumberish[],
      ownerFeePercent: BigNumberish
    ],
    [bigint[]],
    "view"
  >;
  getFunction(
    nameOrSignature: "getRandom"
  ): TypedContractMethod<[a: BigNumberish, b: BigNumberish], [bigint], "view">;
  getFunction(
    nameOrSignature: "getRandomNumbers"
  ): TypedContractMethod<
    [a: BigNumberish, b: BigNumberish, count: BigNumberish],
    [bigint[]],
    "view"
  >;

  filters: {};
}